node-neigh: Introduce dedicated lock
This commit introduces a dedicated lock for managing node neighbor
entries.

Previously, the neighbor subsystem was sharing the same lock with the
node manager. This resulted in the lock starvation when the periodic
neighbor refresh was taking longer than expected and it was queuing up
the refresh requests.

To avoid the starvation which can slow down node updates, we introduce
the dedicated lock. Also, we make sure that insertNeighbor() won't block
NodeUpdate() in the case of the scenario above by running it in a
separate goroutine (same applies to deleteNeighbor() and NodeDelete()).

Signed-off-by: Martynas Pumputis <m@lambda.lt>